requirements.txt
bleach==1.5.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
enum34==1.1.6
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.7.1
h5py==2.7.1
html5lib==0.9999999
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Keras==2.1.2
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.14.0
Pillow==5.0.0
protobuf==3.5.1
PyYAML==3.12
requests==2.18.4
scipy==1.0.0
six==1.11.0
SQLAlchemy==1.2.0
tensorflow==1.4.1
tensorflow-tensorboard==0.4.0rc3
urllib3==1.22
Werkzeug==0.14.1
